requirements.txt
aiosmtplib==5.1.0
alembic==1.12.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
bcrypt==4.0.1
billiard==4.2.4
blinker==1.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
Pillow==10.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
regex==2025.11.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
vine==5.1.0
wcwidth==0.6.0
gunicorn==21.2.0
uvicorn[standard]==0.42.0